requirements.txt
a2a-sdk==0.2.16
absl-py==2.3.1
absolufy-imports==0.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.1
cloudpickle==3.1.1
crashtest==0.4.1
cryptography==45.0.5
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
docstring_parser==0.17.0
dulwich==0.22.8
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
google-adk==1.13.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.110.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.1
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.27.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googlemaps
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.74.0
grpcio-reflection==1.71.2
grpcio-status==1.71.2
grpcio-tools==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
idna==3.10
# Editable Git install (image-scoring==0.1.0) with either a deleted local remote or invalid URI:
# '/usr/local/git/git-google/etc/submodule_dogfood_config'
-e /Users/bharen/code/cloud/Agents/image-scoring/image_scoring
importlib_metadata==8.7.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
litellm==1.74.8
Mako==1.3.10
MarkupSafe==3.0.2
mcp==1.12.2
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.6.3
nltk==3.9.1
numpy==2.3.2
openai==1.97.1
opentelemetry-api==1.35.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
pandas==2.3.1
pbs-installer==2025.7.12
pillow==10.4.0
pkginfo==1.12.1.2
platformdirs==4.3.8
playwright==1.54.0
poetry==2.1.3
poetry-core==2.1.3
poetry-plugin-export==1.9.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rouge_score==0.1.2
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uv==0.7.3
uvicorn==0.35.0
virtualenv==20.31.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xattr==1.1.4
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
